\docType{class}
\name{pop-class}
\alias{pop-class}
\title{Class \code{pop}}
\description{
Base class for the ClusterCall package. 
}
\section{Slots}{

\describe{
\item{\code{theta}}{Marker by sample matrix of theta values}

\item{\code{r}}{Marker by sample matrix of r values (optional)}
}}

\references{
Schmitz Carley et al. (2017) Automated tetraploid genotype calling by hierarchical clustering. Theoretical and Applied Genetics 130:717–726 DOI: 10.1007/s00122-016-2845-5. 
}

\examples{
	showClass("pop")}